fs: add ksys_ftruncate() wrapper; remove in-kernel calls to sys_ftruncate()
Using the ksys_ftruncate() wrapper allows us to get rid of in-kernel
calls to the sys_ftruncate() syscall. The ksys_ prefix denotes that this
function is meant as a drop-in replacement for the syscall. In
particular, it uses the same calling convention as sys_ftruncate().

This patch is part of a series which removes in-kernel calls to syscalls.
On this basis, the syscall entry path can be streamlined. For details, see
http://lkml.kernel.org/r/20180325162527.GA17492@light.dominikbrodowski.net

Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Andrew Morton <akpm@linux-foundation.org>
Change-Id: I5bcdfc0c3b6245a4dfbd08381220d44a59737a99
Signed-off-by: Dominik Brodowski <linux@dominikbrodowski.net>

Conflicts: fs/internal.h
Change-Id: I8cd54a32c4f216548b522da3bc2ab1577cbb89e9
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>