arm64: Fix section mismatch with LTO caused by ambiguous const
Due to how dt_supported_cpu_ops and acpi_supported_cpu_ops are used,
they can be placed in a different section by the compiler when LTO is
used because it thinks that it belongs in another section. To really
make it clear to GCC that these belong in the __initconst section,
make the variables themselves const and refactor cpu_get_ops()
accordingly to make it compile.

Change-Id: I2b6d300c5dc6f69663ac9e790b44e6c3436acb78
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>