drm: Boost DDR bus upon running an atomic ioctl
In order to reduce jank, request a DDR bus boost whenever an atomic
ioctl is sent, which is what our display HAL uses to commit frames to
the display. The boost should be sufficient enough to render the
display's maximum FPS without any dropped frames when there is no
significant external source of load.

Change-Id: I1ee6bcad780441de0d6dfc871763f322715e52b1
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>