fingerprint: gf3208: Rewrite for performance
goodix's implementation is quite generic at it's core and so,
there was room for improvement without hurting the basic functions
debugging and unsupported features are removed and also
removes the need for a display notifier and a wakelock as
our SoC can receive interrupts from sleep and respond just fine.

generally, mutexes and other bogus checks have also been removed
and indents have been fixed, driver should be a lot more readable
now that everything's in one place.

this rewrite can be adapted to work on the majority of goodix sensors
(with minor edits, say, pinctrl or gpio names or new regulators and such.)

tested to be working on 3208, 3258, and fod variants
fp sensor performance had improved by over 50% on 8953 with the same rewrite.

https://github.com/Jebaitedneko/android_kernel_xiaomi_vayu/tree/caf/drivers/input/fingerprint/goodix_ta_vayu
HEAD: 3f98163846bf6250711048f30876a34b99c29b8c

Change-Id: I6df2a712ef07ea191092218dcb53856a2a2afa34
Signed-off-by: Jebaitedneko <Jebaitedneko@gmail.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>