input/fp: gf3208: Set fingerprintd priority to max when screen is off
Give fingerprintd the highest priority (MIN_NICE) when the screen is off to
speed up fingerprint processing, and then reset its priority back to normal
when the screen is on.

There are now 2 fingerprint related processes:
 - android.hardware.biometrics.fingerprint@2.1-service -> comm:fingerprint@2.1
 - com.qualcomm.qti.biometrics.fingerprint.service -> comm:erprint.service

For the match to work properly, use strstr() with "erprint".

[ghostrider-reborn]
- device_available was being set to 0 during fp-unlock-and-wake
which led to niceness being stuck at -20 (except in dt2w)
- Also use xiaomi bridge interface hack to slightly reduce
unlock times, no idea why it was commented in the first place

Change-Id: Ice4a384f99ae4201b2e5e942b0c8cc16c1190f52
Co-authored-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Co-authored-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>