thaw_fingerprintd: Use strncmp() instead of memcmp()
memcmp() requires specific filename, and the fingerprintd hal can be upgraded to a newer version which will require us to modify the code again, hence let's switch to strncmp().

This makes the code more efficient in a way as we're using sizeof() which is determined at compile-time, so the compiler will calculate the size of the string and make it a constant.

Suggested-by: Justin Crawford <Justin@stacksmash.net>
Signed-off-by: Justin Crawford <Justin@stacksmash.net>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Change-Id: I6fe4a3f81a7f845841ffc1d8da6bed018f122a9b
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>