cpufreq: schedutil: Remove magic 1.25 headroom from apply_dvfs_headroom()
Replace 1.25 headroom in apply_dvfs_headroom() with better dynamic
logic.

Instead of the magical 1.25 headroom, use the new approximate_util_avg()
to provide headroom based on the dvfs_update_delay; which is the period
at which the cpufreq governor will send DVFS updates to the hardware.

Add a new percpu dvfs_update_delay that can be cheaply accessed whenever
apply_dvfs_headroom() is called. We expect cpufreq governors that rely
on util to drive its DVFS logic/algorithm to populate these percpu
variables. schedutil is the only such governor at the moment.

The behavior of schedutil will change as the headroom will be less than
1.25 for most systems as the rate_limit_us is usually short.

Signed-off-by: Qais Yousef <qyousef@layalina.io>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>

[Richard Raya: Adapted to 4.14:
 - approximate_util_avg() reimplemented with plain PELT decay math,
   including the accumulate_sum() contribution scaling
 - se.slice (EEVDF) -> sched_slice() (4.14 equivalent; exported for
   schedutil)]

Change-Id: Ie35b72cfcff3eaaadaba54b57c8a0ef21953d0c4
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>